Module Ez_file

module FileAbstract : sig ... end

File operations with filenames represented by an abstract type FileGen.t

module FileChannel : sig ... end

File operations on the stdlib in_channel/out_channel types

module FileDirMaker : sig ... end

module FileOS : sig ... end

Operating-System specific values

module FileSelector : sig ... end

module FileSig : sig ... end

Signatures exported by other modules

module FileString : sig ... end

File operations with filenames represented by the string type

module V1 : sig ... end