Module FileSig

Signatures exported by other modules

module type CONTENT_OPERATIONS = sig ... end

This module provides functions to read and write complete files from and to strings, or working on file lines.

module type DIRECTORY_OPERATIONS = sig ... end

This modules provides functions to create directories, read, iter or remove directories, recursively or not.

module type FILENAME_OPERATIONS = sig ... end

module type FILE_OPERATIONS = sig ... end